learn_dspy/pyproject.toml
[project]
name = "learn-dspy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "ddgs>=9.10.0",
    "dspy>=3.1.2",
    "dspy-qdrant>=0.1.4",
    "duckduckgo-search>=8.1.1",
    "faiss-cpu>=1.13.2",
    "jq>=1.11.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.10",
    "mem0ai>=0.1.115",
    "mlflow>=2.9.0",
    "neo4j>=6.1.0",
    "pandas>=2.2,<3",
    "protobuf<4",
    "python-dotenv>=1.2.1",
]
